separate TreeReader into .h and .cc files
